A lot of changes in the works. Wanted to push early though in case you looked at it and started working on a foundation that I removed. By the way, I left TODOs in the places that need to be done, or looked into later.
Currently, the idea I'm trying to implement is this.
You are advertising betting randomly or maybe to the center patch depending on the button on the interface. So you can spread betting by advertisements or by word of mouth by the people that just received that advertising. However, whenever people meet someone that has bet or hears from someone that has bet, they are invariably influenced by that person's success or failure. So each turtle also has a "influence" that is based on the advertisers "success". So then this influence measure will be applied onto the turtles current "likelihood" measure and you make bets accordingly.

Next steps after implementing this:
1) Only let people that have been advertised to, or already know of the betting exists to be able to bet on every draw.
2) Look into some negative influences...?
3) Keep updating the graphical interface